Restructure all the packages and untangle a lot of dependencies
* That also means that massive improvements were made to the API. For
example, there's now a widget interface instead of a class, which solves
a lot of problems users and even the API itself had previously.
* Moreover, all widgets are now described through interfaces and then
implemented by concrete classes. That adds an additional layer of
possible abstraction if you need it -- and in many cases, you actually
do.
* Another major improvement is the new and way more intuitive & flexible
tooltip system.
* A lot of other, smaller improvements were made while the refactoring
went on.