Add external classes which take care of recursion processes like drawing
The event handlers on individual containers do no longer go about
calling event handlers on their children! Instead, they implement
WidgetParent and provide a list of their child widgets as well as an
appropriate sub-viewport for each child through that interface.

The reason for this change is quite simple. Many of the recursions --
especially the revalidation one -- became so complex that the amount of
rules each parent widget had to implement exploded. By externalizing all
that complicated logic into recursion classes, the widget classes
themselves are reduced to the actual widget logic. And of course, we no
longer duplicate the code multiple times in multiple locations.