Add support for nested focus
Imagine a popup that opens whenever a certain widget is focused (just
like a dropdown menu). Now imagine that this popup has some focusable
widgets to, e.g. a text field. With the new support for nested focus,
the popup and the text field inside the popup can be focused at the same
time since they lie on the same _path_ through the widget tree.